// src/types/firebase-errors.ts
export type FirebaseErrorCode =
  | "auth/user-not-found"
  | "auth/wrong-password"
  | "auth/invalid-email"
  | "auth/user-disabled"
  | "auth/email-already-in-use"
  | "auth/operation-not-allowed"
  | "auth/weak-password"
  | "auth/popup-blocked"
  | "auth/popup-closed-by-user"
  | "auth/cancelled-popup-request"
  | "auth/invalid-credential"
  | "auth/account-exists-with-different-credential"
  | "auth/network-request-failed"
  | "auth/too-many-requests"
  | "auth/requires-recent-login"
  | "auth/user-token-expired"
  | "auth/web-storage-unsupported"
  | "auth/unverified-email"
  | "auth/invalid-verification-code"
  | "auth/invalid-verification-id"
  | "auth/captcha-check-failed"
  | string;